segregate types in the order and response modules
most types used in Order are in order.rs
most types used in CommandResponse are in response.rs
types relevant to TLS certificates are in certificate.rs
