+++ bin/src/ctl/display.rs
@@ -6,12 +6,10 @@ use std::{
 use anyhow::{self, bail, Context};
 use prettytable::{Row, Table};
 
-use sozu_command_lib::{
-    order::{CommandResponseContent, ListedFrontends, ListenersList, WorkerInfo},
-    worker::{
-        AggregatedMetricsData, ClusterMetricsData, FilteredData, QueryAnswer,
-        QueryAnswerCertificate, QueryAnswerMetrics, Route, WorkerMetrics,
-    },
+use sozu_command_lib::response::{
+    AggregatedMetricsData, ClusterMetricsData, CommandResponseContent, FilteredData,
+    ListedFrontends, ListenersList, QueryAnswer, QueryAnswerCertificate, QueryAnswerMetrics, Route,
+    WorkerInfo, WorkerMetrics,
 };
 
 pub fn print_listeners(listeners_list: ListenersList) {
