+++ bin/src/upgrade.rs
@@ -16,7 +16,7 @@ use serde::{Deserialize, Serialize};
 use tempfile::tempfile;
 
 use sozu_command_lib::{
-    channel::Channel, config::Config, order::RunState, state::ConfigState, worker::InnerOrder,
+    channel::Channel, config::Config, order::InnerOrder, response::RunState, state::ConfigState,
 };
 
 use crate::{
