+++ command/src/config.rs
@@ -13,13 +13,15 @@ use anyhow::{bail, Context};
 use toml;
 
 use crate::{
-    certificate::split_certificate_chain,
-    order::Order,
-    worker::{
-        ActivateListener, AddCertificate, Backend, CertificateAndKey, Cluster, HttpFrontend,
-        HttpListenerConfig, HttpsListenerConfig, InnerOrder, ListenerType, LoadBalancingAlgorithms,
-        LoadBalancingParams, LoadMetric, PathRule, Route, RulePosition, TcpFrontend,
-        TcpListenerConfig, TlsVersion,
+    certificate::TlsVersion,
+    order::{ActivateListener, Cluster, InnerOrder, ListenerType, LoadBalancingParams},
+    response::{Backend, HttpListenerConfig, TcpFrontend},
+};
+use crate::{
+    certificate::{split_certificate_chain, CertificateAndKey},
+    order::{AddCertificate, LoadBalancingAlgorithms, LoadMetric, Order},
+    response::{
+        HttpFrontend, HttpsListenerConfig, PathRule, Route, RulePosition, TcpListenerConfig,
     },
 };
