+++ command/src/lib.rs
@@ -10,6 +10,7 @@ pub mod config;
 pub mod order;
 pub mod parser;
 pub mod ready;
+pub mod response;
 pub mod scm_socket;
 pub mod state;
 pub mod worker;
